fix: permission checks moved to server components for roles and newsletter
